# Performance: Add TimerWheel to Bulk to improve latency